Moving error aggregator to vterrors.
It is generic enough.
